requirements.txt
python-dotenv
requests
matplotlib
fastapi[standard]